// src/enums/websocket-events.ts
export enum EventType {
  FRIENDSHIP_REQUEST = "FRIENDSHIP_REQUEST",
  FRIENDSHIP_ACCEPTED = "FRIENDSHIP_ACCEPTED",
  USER_JOIN_ROOM = "USER_JOIN_ROOM",
  MESSAGE_SENT = "MESSAGE_SENT",
  MESSAGE_RECEIVED = "MESSAGE_RECEIVED",
  USER_LOGGED_IN = "USER_LOGGED_IN",
  FRIEND_LOGGED_IN = "FRIEND_LOGGED_IN",
  USER_LOGGED_OUT = "USER_LOGGED_OUT",
  FRIEND_LOGGED_OUT = "FRIEND_LOGGED_OUT",
  USER_TYPING = "USER_TYPING",
  USER_TYPING_STOPPED = "USER_TYPING_STOPPED",
  USER_NAVIGATED_AWAY = "USER_NAVIGATED_AWAY",
  MESSAGE_RECEIVED_NOTIFICATION = "USER_NAVIGATED_AWAY",
}